all responsive except factory, made splash page
