+++ fashoid/pages/learnmore.js
@@ -31,26 +31,31 @@ const VideoCont = styled.iframe`
   box-shadow: 2px 2px 5px rgb(31 31 31/0.6);
   animation: ${Drop} 3s;
 `
+const LearnMoreCont = styled.div`
+  animation: ${SlideRightDiff} 1.5s;
+`
+const NavBtns = styled.div`
+  width: 74vw;
+  display: flex;
+  padding: 20px 0px 0px 0px;
+  justify-content: space-between;
+  align-items: center;
+`
 const Home = styled(HomeIcon)`
     color: #333;
     height: 40px;
     width: auto;
-    position: absolute;
-    margin: 15px 0px 0px 330px;
+    position: relative;
     cursor: pointer;
     animation: ${topToBottom} 1s;
 `
 const Back = styled(ArrowBackIosSharpIcon)`
-    position: absolute;
-    margin: 20px 320px 0px 0px;
+    position: relative;
     height: 30px;
     width: auto;
     color: #333;
     cursor: pointer;
     animation: ${topToBottom} 1s;
-`
-const LearnMoreCont = styled.div`
-  animation: ${SlideRightDiff} 1.5s;
 `
 
 export default function LearnMore() {
@@ -58,16 +63,18 @@ export default function LearnMore() {
   const r = useRouter();
 
   return <MainCont>
-    <Home
-      onClick={
-        () => r.push("/")
-      }
-    />
-    <Back
-      onClick={
-        () => r.back("")
-      }
-    />
+    <NavBtns>
+      <Back
+        onClick={
+          () => r.back("")
+        }
+      />
+      <Home
+        onClick={
+          () => r.push("/home")
+        }
+      />
+    </NavBtns>
     <Background img="/Home_bg.svg" />
     <LearnMoreCont>
       <Img img="/learnmoreguy.svg" w="290" h="230" />
